pyproject.toml
[project]
name = "babysleep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]

[tool.scripts]
main = "main"